feat(invite): share InviteCodeBody and render the Figma QR card
Extract the Scan/Invite toggle plus its Invite/Scan tabs into
InviteCodeBody, the single shared implementation, and reduce
InviteCodeOverlay to a thin composer that drops InviteCodeBody under
its floating liquid-glass nav. Both the full-screen overlay and the
embedded new-convo flow now use the same body, so the toggle and tabs
do not fork.

Restore the Figma QR card around the legacy glyph: the Invite tab's QR
sits on a fillSubtle (#F5F5F5) card with corner radius extraLarge (56)
and roomy padding from the card edge, while the glyph itself stays the
legacy QRCodeGenerator (Q error correction, rounded modules, bigger
pixels) with the center conversation avatar (avatar -> emoji ->
monogram). The QR quiet zone matches the card fill.
